update UI tests: close sparkle alerts
