+++ SlimHUDUITests/AboutUITest.swift
@@ -10,41 +10,23 @@
 import XCTest
 
 final class AboutUITests: XCTestCase {
-    let app = XCUIApplication()
-    
-    override func setUpWithError() throws {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-
-        // In UI tests it is usually best to stop immediately when a failure occurs.
-        continueAfterFailure = true
-        app.launch()
-
-        // In UI tests it’s important to set the initial state - such as interface orientation - required for your tests before they run. The setUp method is a good place to do this.
-    }
-    
-    
-    func testAppInMenuBar() throws {
-        
-        let menuBarsQuery = app.menuBars
-        let statusItem = menuBarsQuery.children(matching: .statusItem).element(boundBy: 0)
-        
-        XCTAssert(statusItem.waitForExistence(timeout: 5))
-        
-        let actualStatusItemScreenshot = statusItem.screenshot()
-        let attachment = XCTAttachment(screenshot: actualStatusItemScreenshot)
-        attachment.name = "Status Item screen"
-        attachment.lifetime = .keepAlways
-        add(attachment)
-    }
-    
     func testOpenAboutWindow() throws {
+        let app = XCUIApplication()
+        app.launch()
         
         let menuBarsQuery = app.menuBars
         let statusItem = menuBarsQuery.children(matching: .statusItem).element(boundBy: 0)
         
         XCTAssert(statusItem.waitForExistence(timeout: 5))
         statusItem.click()
-
+        
+        
+        // Close Alert opened by Sparkle
+        SparkleUITests.closeAlerts(app: app)
+        // Click again on about menu item, in case some alerts where closed
+        statusItem.click()
+        
+        
         let aboutMenuItem = menuBarsQuery.menuItems["About..."]
         XCTAssert(aboutMenuItem.waitForExistence(timeout: 5))
         aboutMenuItem.click()
