+++ SlimHUDUITests/SlimHUDUITestsLaunchTests.swift
@@ -8,30 +8,7 @@
 
 import XCTest
 
-final class SlimHUDUITestsLaunchTests: XCTestCase {
-
-    override class var runsForEachTargetApplicationUIConfiguration: Bool {
-        false
-    }
-    
-    override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-    }
-
-    override func setUpWithError() throws {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-
-        // In UI tests it is usually best to stop immediately when a failure occurs.
-        continueAfterFailure = false
-
-        // In UI tests it’s important to set the initial state - such as interface orientation - required for your tests before they run. The setUp method is a good place to do this.
-    }
-    
-    func testLaunch() throws {
-        let app = XCUIApplication()
-        app.launch()
-    }
-    
+final class SlimHUDUITestsLaunchTests: XCTestCase {   
     func testAppInMenuBar() throws {
         let app = XCUIApplication()
         app.launch()
