+++ SlimHUDUITests/SparkleUITests/SparkleUITest.swift
@@ -1,0 +1,17 @@
+//
+//  SparkleUITest.swift
+//  SlimHUDUITests
+//
+//  Created by Alex Perathoner on 23/12/22.
+//  Copyright © 2022 Alex Perathoner. All rights reserved.
+//
+
+import XCTest
+
+final class SparkleUITests: XCTestCase {
+    static public func closeAlerts(app: XCUIApplication) {
+        while(app.dialogs.count > 0) {
+            app.dialogs.firstMatch.buttons.firstMatch.doubleClick()
+        }
+    }
+}
